market-dashboard/requirements.txt
psycopg2-binary==2.9.10
akshare==1.18.55
pandas==2.3.3
openpyxl==3.1.5
fastapi==0.124.2
uvicorn[standard]==0.38.0
python-multipart==0.0.20
httpx==0.28.1
cryptography==46.0.3
